* Added build configurations for ARM and made a few minor changes to the source to make it build for ARM.  All of the libraries except the unit tests build for ARM (I'm not yet sure what will be required to make them build; we'll tackle that when we actually get an ARM device on which we can test).
 * Fixed a couple of 64-bit narrowing conversion warnings.  The Solution now builds cleanly in all twelve configurations.

+++ utility/create_sln_from_directories/main.cpp
@@ -156,7 +156,7 @@ namespace
         std::copy(begin(result), begin(result) + sizeof guid, reinterpret_cast<byte_iterator>(&guid));
 
         std::vector<wchar_t> guid_string(40);
-        if (::StringFromGUID2(guid, guid_string.data(), guid_string.size()) == 0)
+        if (::StringFromGUID2(guid, guid_string.data(), static_cast<int>(guid_string.size())) == 0)
             throw std::runtime_error("failed to stringify guid");
 
         std::vector<char> narrow_guid_string(begin(guid_string), end(guid_string));
